Feature/accountancy 2018/fixed assets contextual update

+++ app/services/fixed_asset/transitions/sell.rb
@@ -1,6 +1,8 @@
 class FixedAsset
   module Transitions
     class Sell < Transitionable::Transition
+      include Depreciable
+
       event :sell
       from :in_use
       to :sold
@@ -14,19 +16,20 @@ class FixedAsset
       def transition
         resource.sold_on ||= @sold_on
         resource.transaction do
-          active = resource.depreciations.on @sold_on
-          split_depreciation! active, @sold_on unless active.stopped_on == @sold_on
+          active = resource.depreciation_on @sold_on
+          split_depreciation! active, @sold_on if active && @sold_on < active.stopped_on
 
           resource.depreciations.up_to(@sold_on).each { |d| d.update! accountable: true }
 
           resource.update! state: :sold
 
-          resource.depreciations.following(active).destroy_all
+          resource.depreciations.following(active).destroy_all if active
 
           resource.depreciations.update_all locked: true
 
           resource.sale.invoice unless resource.sale.invoice?
 
+          resource.product.reload
           resource.product.update! dead_at: @sold_on
           true
         end
@@ -34,16 +37,11 @@ class FixedAsset
 
       def can_run?
         super && resource.valid? &&
-          sold_on_during_opened_financial_year &&
+          FinancialYear.on(@sold_on)&.opened? &&
+          depreciations_valid?(@sold_on) &&
           resource.product &&
           resource.sale
       end
-
-      private
-
-        def sold_on_during_opened_financial_year
-          FinancialYear.on(@sold_on)&.opened?
-        end
     end
   end
 end
